+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.11.2] - 2023-01-03
 ### Added
 - insert: `with_timeouts` to manage timeouts.
 - inserter: `with_timeouts` and `set_timeouts` to manage timeouts.
@@ -219,7 +221,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client::query()` for selecting from tables and DDL statements.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.11.1...HEAD
+[Unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.11.2...HEAD
+[0.11.2]: https://github.com/loyd/clickhouse.rs/compare/v0.11.1...v0.11.2
 [0.11.1]: https://github.com/loyd/clickhouse.rs/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/loyd/clickhouse.rs/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/loyd/clickhouse.rs/compare/v0.9.3...v0.10.0